Move TableDefinitionsRepository to runner package
