+++ tempto-core/src/main/java/com/teradata/tempto/fulfillment/table/hive/tpch/TpchTableDefinitions.java
@@ -14,8 +14,8 @@
 
 package com.teradata.tempto.fulfillment.table.hive.tpch;
 
+import com.teradata.tempto.fulfillment.table.RepositoryTableDefinition;
 import com.teradata.tempto.fulfillment.table.hive.HiveTableDefinition;
-import com.teradata.tempto.fulfillment.table.TableDefinitionsRepository.RepositoryTableDefinition;
 
 // Table definitions according to: http://www.tpc.org/tpc_documents_current_versions/pdf/tpc-h_v2.17.1.pdf
 // TODO: support for CHAR
